feat: did-exchange REST endpoints to user controller
- DIDExchange REST operations to use controller commands
- Removed `pkg/restapi` and all restapi initialization and opts moved to
`pkg/controller/restapi`
